requirements.txt
onnxruntime-directml
insightface
opencv-python
numpy
psutil
packaging
pyinstaller